--- a/codegen/templates/object.gotpl
+++ b/codegen/templates/object.gotpl
@@ -32,37 +32,39 @@ func (ec *executionContext) _{{$object.GQLType}}(ctx context.Context, sel ast.Se
 		})
 	{{end}}
 
-	{{if $object.IsConcurrent}} var wg sync.WaitGroup {{end}}
-	out := graphql.NewOrderedMap(len(fields))
+	out := graphql.NewFieldSet(fields)
 	invalid := false
 	for i, field := range fields {
-		out.Keys[i] = field.Alias
-
 		switch field.Name {
 		case "__typename":
 			out.Values[i] = graphql.MarshalString({{$object.GQLType|quote}})
 		{{- range $field := $object.Fields }}
 		case "{{$field.GQLName}}":
 			{{- if $field.IsConcurrent }}
-				wg.Add(1)
-				go func(i int, field graphql.CollectedField) {
-			{{- end }}
+				field := field
+				out.Concurrently(i, func() (res graphql.Marshaler) {
+					res = ec._{{$object.GQLType}}_{{$field.GQLName}}(ctx, field{{if not $object.Root}}, obj{{end}})
+					{{- if $field.ASTType.NonNull }}
+						if res == graphql.Null {
+							invalid = true
+						}
+					{{- end }}
+					return res
+				})
+			{{- else }}
 				out.Values[i] = ec._{{$object.GQLType}}_{{$field.GQLName}}(ctx, field{{if not $object.Root}}, obj{{end}})
 				{{- if $field.ASTType.NonNull }}
 					if out.Values[i] == graphql.Null {
 						invalid = true
 					}
 				{{- end }}
-			{{- if $field.IsConcurrent }}
-					wg.Done()
-				}(i, field)
 			{{- end }}
 		{{- end }}
 		default:
 			panic("unknown field " + strconv.Quote(field.Name))
 		}
 	}
-	{{if $object.IsConcurrent}} wg.Wait() {{end}}
+	out.Dispatch()
 	if invalid { return graphql.Null }
 	return out
 }