--- a/graphql/jsonw.go
+++ b/graphql/jsonw.go
@@ -26,40 +26,10 @@ type Unmarshaler interface {
 	UnmarshalGQL(v interface{}) error
 }
 
-type OrderedMap struct {
-	Keys   []string
-	Values []Marshaler
-}
-
 type WriterFunc func(writer io.Writer)
 
 func (f WriterFunc) MarshalGQL(w io.Writer) {
 	f(w)
-}
-
-func NewOrderedMap(len int) *OrderedMap {
-	return &OrderedMap{
-		Keys:   make([]string, len),
-		Values: make([]Marshaler, len),
-	}
-}
-
-func (m *OrderedMap) Add(key string, value Marshaler) {
-	m.Keys = append(m.Keys, key)
-	m.Values = append(m.Values, value)
-}
-
-func (m *OrderedMap) MarshalGQL(writer io.Writer) {
-	writer.Write(openBrace)
-	for i, key := range m.Keys {
-		if i != 0 {
-			writer.Write(comma)
-		}
-		writeQuotedString(writer, key)
-		writer.Write(colon)
-		m.Values[i].MarshalGQL(writer)
-	}
-	writer.Write(closeBrace)
 }
 
 type Array []Marshaler