Added a benchmark for a simple query and make a 2 quick optimisations:
 - don't use strconv.Quote when writing object keys to save a bunch of allocs
 - execute the first concurrent field on the request goroutine, instead always creating a new goroutine.

before
```
    go test -benchtime=5s -bench=. -benchmem
    goos: linux
    goarch: amd64
    pkg: github.com/99designs/gqlgen/example/starwars
    BenchmarkSimpleQueryNoArgs-8      200000             32680 ns/op            6357 B/op        126 allocs/op
    PASS
    ok      github.com/99designs/gqlgen/example/starwars    9.901s
```
![image](https://user-images.githubusercontent.com/2247982/49680996-14545900-faef-11e8-9cd9-26c138930370.png)

You can see `morestack` dominating the graph here  (28%). This happens because we create new goroutines around every resolver, and even though go does recycle old goroutine in a pool, it doesn't maintain their stacks.

after
```
    go test -benchtime=5s -bench=. -benchmem    
    goos: linux
    goarch: amd64
    pkg: github.com/99designs/gqlgen/example/starwars
    BenchmarkSimpleQueryNoArgs-8      300000             25093 ns/op            6453 B/op        114 allocs/op
    PASS
    ok      github.com/99designs/gqlgen/example/starwars    10.807s
```
![image](https://user-images.githubusercontent.com/2247982/49680908-31d4f300-faee-11e8-9a9c-5c3256e48b65.png)

~ 30% faster.



Ideas for future tweaks:
 - 15% is still spent on `morestack` + 13% in starwars.ec_Human_starship: Marshalling arrays in results also happens concurrently, theres another high fanout go func() in here. This causes lots of short lived goroutines, which allocate more stack. Probably the next target.
 - 8% in json.Marshal:  This can be replaced with our internal json writer, avoiding all of the reflection and many allocations. Needs a little bit of thought around encoding extensions, do we fall back to json.Marshal?
 - 7% in json.Decode: Maybe there is a faster json decoder out there for this?